+++ ChangeLog.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- NEW : Ajout de la class TechATM pour l'affichage de la page "A propos" *10/05/2022* 2.2.0
 - NEW : Reset proposal dates on archiving - 2.1.0 - *04/10/2021*  
   Need hidden conf PROPALEHISTORY_ARCHIVE_AND_RESET_DATES set to 1
   by activating this conf, reset proposal date and end validity date with today date if above const is enabled